apps/requirements.txt
celery==5.4.0
channels[daphne]==4.2.0
channels-redis==4.2.1
django-celery-results==2.5.1
django-environ==0.12.0
mtcnn==1.0.0
opencv-python==4.11.0.86
pillow==11.1.0
psycopg==3.2.6
redis==5.2.1
tensorflow_cpu==2.19.0
